+++ example/lib/widgets/responsive_widget.dart
@@ -13,16 +13,16 @@ class ResponsiveWidget extends StatelessWidget {
   final Widget? smallScreen;
 
   static bool isSmallScreen(BuildContext context) {
-    return MediaQuery.of(context).size.width < 800;
+    return MediaQuery.sizeOf(context).width < 800;
   }
 
   static bool isLargeScreen(BuildContext context) {
-    return MediaQuery.of(context).size.width > 1200;
+    return MediaQuery.sizeOf(context).width > 1200;
   }
 
   static bool isMediumScreen(BuildContext context) {
-    return MediaQuery.of(context).size.width >= 800 &&
-        MediaQuery.of(context).size.width <= 1200;
+    return MediaQuery.sizeOf(context).width >= 800 &&
+        MediaQuery.sizeOf(context).width <= 1200;
   }
 
   @override