+++ lib/src/models/documents/attribute.dart
@@ -102,6 +102,8 @@ class Attribute<T> {
 
   static final ScriptAttribute script = ScriptAttribute(null);
 
+  // TODO: You might want to mark those as key (mobileWidthKey)
+  // because it was not very clear to a developer that is new to this project
   static const String mobileWidth = 'mobileWidth';
 
   static const String mobileHeight = 'mobileHeight';
@@ -109,6 +111,12 @@ class Attribute<T> {
   static const String mobileMargin = 'mobileMargin';
 
   static const String mobileAlignment = 'mobileAlignment';
+
+  /// For other platforms, for mobile use [mobileAlignment]
+  static const String alignment = 'alignment';
+
+  /// For other platforms, for mobile use [mobileMargin]
+  static const String margin = 'margin';
 
   static const ImageAttribute image = ImageAttribute(null);