+++ lib/src/utils/platform.dart
@@ -1,12 +1,15 @@
 import 'package:device_info_plus/device_info_plus.dart';
-import 'package:flutter/foundation.dart';
+import 'package:flutter/foundation.dart'
+    show kIsWeb, TargetPlatform, defaultTargetPlatform;
 
 bool isMobile([TargetPlatform? targetPlatform]) {
+  if (kIsWeb) return false;
   targetPlatform ??= defaultTargetPlatform;
   return {TargetPlatform.iOS, TargetPlatform.android}.contains(targetPlatform);
 }
 
 bool isDesktop([TargetPlatform? targetPlatform]) {
+  if (kIsWeb) return false;
   targetPlatform ??= defaultTargetPlatform;
   return {TargetPlatform.macOS, TargetPlatform.linux, TargetPlatform.windows}
       .contains(targetPlatform);
@@ -18,6 +21,7 @@ bool isKeyboardOS([TargetPlatform? targetPlatform]) {
 }
 
 bool isAppleOS([TargetPlatform? targetPlatform]) {
+  if (kIsWeb) return false;
   targetPlatform ??= defaultTargetPlatform;
   return {
     TargetPlatform.macOS,